requirements.txt
APScheduler==3.9.1.post1
paho-mqtt==1.6.1
pyserial==3.5
python-dotenv==0.21.0
